requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
playwright==1.50.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-slugify==8.0.4
requests==2.32.3
setuptools==75.8.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
